pyproject.toml
[project]
name = "style_transfer"
version = "0.1.0"
description = "A modern neural style transfer application using PyTorch"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.66.0",
    "matplotlib>=3.8.0",
    "pyside6>=6.7.0",
]
requires-python = ">=3.10"

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.1",
    "pytest-timeout>=2.1.0",
    "pytest-mock>=3.11.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--verbose --tb=short --cov=style_transfer --cov-report=term-missing --cov-report=html --no-cov-on-fail"